requirements.txt
streamlit
pandas
tweepy
python-dotenv
transformers
torch
emoji
matplotlib
redis
google-api-python-client

